add service interfaces add first test on mvc
